# docs/openapi/novel-platform.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Novel Platform API",
    "version": "0.1.0",
    "description": "Versioned REST contract for the novel platform backend. Browser clients use the same-origin Next BFF: replace the backend prefix `/api/v1` with `/api/novel` for the `public`, `account`, `author`, and `admin` route groups. The BFF owns browser cookies, CSRF validation, and the upstream internal headers; do not call protected backend routes from browser JavaScript directly. `/api/v1/auth/*` is an internal BFF-to-backend contract and is intentionally not exposed by the catch-all browser proxy. This contract describes implemented local platform behavior only. Phone/email verification, WeChat/QQ OAuth, external payment, and production Qwen credentials are not represented as available API integrations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Backend service. Protected and auth routes require BFF-only headers."
    }
  ],
...
  "paths": {
...
    "/api/v1/account/profile": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getAccountProfile",
        "summary": "Get the current account profile",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/AccountProfile",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "put": {
        "tags": ["Reader account"],
        "operationId": "updateAccountProfile",
        "summary": "Update the current account display name",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/AccountProfile",
        "requestBody": { "$ref": "#/components/requestBodies/ProfileUpdateRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v1/account/entitlements": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getAccountEntitlements",
        "summary": "Get current membership and book entitlements",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/AccountEntitlements",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v1/account/bookshelf": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getBookshelf",
        "summary": "Get the current reader bookshelf",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/BookList",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v1/account/bookshelf/{bookId}": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "toggleBookshelfBook",
        "summary": "Add or remove one published book from the current bookshelf",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/BookshelfToggleResult",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/account/checkin": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "checkIn",
        "summary": "Claim the daily fixed 10-point check-in award",
        "description": "The current implementation awards 10 points once per Asia/Shanghai calendar day; it does not expose configurable membership or ticket rules.",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/CheckInResult",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/account/redeem": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "redeemCode",
        "summary": "Redeem one platform code",
        "description": "This is a platform redemption flow, not an external payment gateway.",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/RedeemResult",
        "requestBody": { "$ref": "#/components/requestBodies/RedeemRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/account/wallet": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getWallet",
        "summary": "Get current points and platform tokens",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/Wallet",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v1/account/preferences/reading": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getReadingPreference",
        "summary": "Get the current reading preference",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/ReadingPreference",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "put": {
        "tags": ["Reader account"],
        "operationId": "updateReadingPreference",
        "summary": "Save the current reading preference",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/ReadingPreference",
        "requestBody": { "$ref": "#/components/requestBodies/ReadingPreferenceRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/v1/account/progress": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getReadingProgress",
        "summary": "Get the current reader progress list",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/ReadingProgressList",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "put": {
        "tags": ["Reader account"],
        "operationId": "saveReadingProgress",
        "summary": "Save reading progress for a published chapter",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/ReadingProgress",
        "requestBody": { "$ref": "#/components/requestBodies/ProgressRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/account/books/{bookId}/bookmarks": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getBookmarks",
        "summary": "Get the current reader bookmarks for one book",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/BookmarkList",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "post": {
        "tags": ["Reader account"],
        "operationId": "createBookmark",
        "summary": "Create a reader bookmark and matching reading progress",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/Bookmark",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "requestBody": { "$ref": "#/components/requestBodies/BookmarkRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/account/books/{bookId}/comments": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "createComment",
        "summary": "Create a comment subject to automated/manual review",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/Comment",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "requestBody": { "$ref": "#/components/requestBodies/CommentRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/account/books/{bookId}/rating": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "rateBook",
        "summary": "Set the current reader rating for one book",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/RatingResult",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "requestBody": { "$ref": "#/components/requestBodies/RatingRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/account/books/{bookId}/votes/{type}": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "voteForBook",
        "summary": "Record a current-reader book vote",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/VoteResult",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/VoteType" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/account/books/{bookId}/reward": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "rewardBook",
        "summary": "Reward a book with platform tokens",
        "description": "`Idempotency-Key` is required, is scoped to this reward operation, and must be 128 characters or fewer. Amounts are platform tokens only, not fiat currency.",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/RewardResult",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "$ref": "#/components/requestBodies/AmountRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/account/books/{bookId}/purchase": {
      "post": {
        "tags": ["Reader account"],
        "operationId": "purchaseBook",
        "summary": "Purchase one book with its server-owned platform-token price",
        "description": "The optional request body is ignored; clients cannot set a price. This is not an external payment-gateway API.",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/PurchaseResult",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/account/author-applications": {
      "get": {
        "tags": ["Reader account"],
        "operationId": "getCurrentAuthorApplication",
        "summary": "Get the current account author application, when present",
        "security": [{ "BffSession": [] }],
        "x-response-data-schema": "#/components/schemas/NullableAuthorApplication",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "post": {
        "tags": ["Reader account"],
        "operationId": "applyForAuthorRole",
        "summary": "Submit an author application",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-response-data-schema": "#/components/schemas/AuthorApplication",
        "requestBody": { "$ref": "#/components/requestBodies/AuthorApplicationRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "novel_session",
        "description": "Opaque browser BFF session cookie. The browser never receives the backend session id."
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-CSRF",
        "description": "Required by the BFF on unsafe browser requests and must match its CSRF cookie/session binding."
      },
      "InternalBffKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Internal-Key",
        "description": "Private BFF-to-backend credential. Never send this header from a browser."
      },
      "BackendBffSession": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Bff-Session",
        "description": "Server-side backend session credential held by the BFF only. Never expose it to browser code."
      }
    },
    "parameters": {
...
      "PathBookId": { "name": "bookId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 128 } },
      "VoteType": { "name": "type", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Vote type accepted by the current backend implementation." },
...
    },
    "requestBodies": {
...
      "ProfileUpdateRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileUpdateRequest" } } } },
      "ReadingPreferenceRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReadingPreference" } } } },
      "ProgressRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgressRequest" } } } },
      "BookmarkRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BookmarkRequest" } } } },
      "CommentRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommentRequest" } } } },
      "RatingRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RatingRequest" } } } },
      "AmountRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AmountRequest" } } } },
      "AuthorApplicationRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthorApplicationRequest" } } } },
...
      "RedeemRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RedeemRequest" } } } },
...
    },
    "responses": {
      "Success": { "description": "Successful platform envelope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiEnvelope" } } } },
      "BadRequest": { "description": "Invalid request envelope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Unauthorized": { "description": "Missing, expired, or invalid BFF session/internal BFF credential.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "NotFound": { "description": "Requested resource does not exist or is not visible to the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Conflict": { "description": "Requested state transition conflicts with current durable state.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer", "example": 200 }, "msg": { "type": "string", "example": "ok" }, "data": {} }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer" }, "msg": { "type": "string" }, "data": { "type": "null" } }
      },
      "Book": {
        "type": "object",
        "required": ["id", "title", "author", "category", "words", "serialStatus", "synopsis", "cover", "status", "authorId", "heat", "purchasePrice"],
        "properties": {
          "id": { "type": "integer", "format": "int64" }, "title": { "type": "string" }, "author": { "type": "string" }, "category": { "type": "string" }, "words": { "type": "integer" }, "serialStatus": { "type": "string" }, "synopsis": { "type": "string" }, "cover": { "type": "string" }, "status": { "type": "string", "enum": ["DRAFT", "PENDING_REVIEW", "PUBLISHED", "NEEDS_REVIEW", "REJECTED", "OFFLINE"] }, "authorId": { "type": "integer", "format": "int64" }, "heat": { "type": "integer", "format": "int64" }, "purchasePrice": { "type": "integer", "format": "int64", "description": "Server-owned platform-token amount." }
        }
      },
      "BookList": { "type": "array", "items": { "$ref": "#/components/schemas/Book" } },
...
      "Comment": { "type": "object", "required": ["id", "bookId", "userId", "authorName", "content", "status", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": ["integer", "null"], "format": "int64" }, "userId": { "type": "integer", "format": "int64" }, "authorName": { "type": "string" }, "content": { "type": "string" }, "status": { "type": "string", "enum": ["PENDING_REVIEW", "VISIBLE", "REJECTED"] }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "AccountProfile": { "type": "object", "required": ["id", "name", "roles"], "properties": { "id": { "type": "integer", "format": "int64" }, "name": { "type": "string" }, "roles": { "type": "array", "items": { "type": "string", "enum": ["READER", "AUTHOR", "ADMIN"] }, "uniqueItems": true } } },
      "AccountEntitlements": { "type": "object", "required": ["membership", "books"], "properties": { "membership": { "oneOf": [{ "type": "null" }, { "$ref": "#/components/schemas/MembershipEntitlement" }] }, "books": { "type": "array", "items": { "$ref": "#/components/schemas/BookEntitlement" } } } },
      "MembershipEntitlement": { "type": "object", "required": ["expiresAt", "active"], "properties": { "expiresAt": { "type": "string", "format": "date-time" }, "active": { "type": "boolean" } } },
      "BookEntitlement": { "type": "object", "required": ["bookId", "bookTitle", "sourceType", "sourceReference", "purchaseAmount", "amountUnit", "acquiredAt"], "properties": { "bookId": { "type": "integer", "format": "int64" }, "bookTitle": { "type": "string" }, "sourceType": { "type": "string" }, "sourceReference": { "type": "string" }, "purchaseAmount": { "type": "integer", "format": "int64" }, "amountUnit": { "type": "string", "const": "TOKEN" }, "acquiredAt": { "type": "string", "format": "date-time" } } },
      "ReadingPreference": { "type": "object", "required": ["theme", "font", "fontSize", "lineHeight", "brightness", "pageMode"], "properties": { "theme": { "type": "string", "enum": ["paper", "night", "sepia"] }, "font": { "type": "string", "minLength": 1 }, "fontSize": { "type": "integer", "minimum": 14, "maximum": 32 }, "lineHeight": { "type": "integer", "minimum": 120, "maximum": 260 }, "brightness": { "type": "integer", "minimum": 10, "maximum": 100 }, "pageMode": { "type": "string", "enum": ["slide", "cover", "simulation"] } } },
      "ReadingProgress": { "type": "object", "required": ["bookId", "chapterId", "offset", "updatedAt"], "properties": { "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": "integer", "format": "int64" }, "offset": { "type": "integer", "minimum": 0 }, "updatedAt": { "type": "string", "format": "date-time" } } },
      "ReadingProgressList": { "type": "array", "items": { "$ref": "#/components/schemas/ReadingProgress" } },
      "Bookmark": { "type": "object", "required": ["id", "bookId", "chapterId", "offset", "note", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": "integer", "format": "int64" }, "offset": { "type": "integer", "minimum": 0 }, "note": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" } } },
      "BookmarkList": { "type": "array", "items": { "$ref": "#/components/schemas/Bookmark" } },
      "Wallet": { "type": "object", "required": ["points", "tokens"], "properties": { "points": { "type": "integer", "format": "int64" }, "tokens": { "type": "integer", "format": "int64" } } },
      "BookshelfToggleResult": { "type": "object", "required": ["saved"], "properties": { "saved": { "type": "boolean" } } },
      "CheckInResult": { "type": "object", "required": ["points", "awarded"], "properties": { "points": { "type": "integer", "format": "int64" }, "awarded": { "type": "integer", "const": 10 } } },
      "RedeemResult": { "type": "object", "description": "Current redemption result fields are benefit-dependent.", "additionalProperties": true },
      "RatingResult": { "type": "object", "required": ["average"], "properties": { "average": { "type": "number", "format": "double" } } },
      "VoteResult": { "type": "object", "additionalProperties": true },
      "RewardResult": { "type": "object", "additionalProperties": true, "description": "Reward result includes server-derived balance/state fields." },
      "PurchaseResult": { "type": "object", "additionalProperties": true, "description": "Purchase result includes server-derived entitlement/balance state." },
      "AuthorApplication": { "type": "object", "required": ["id", "userId", "penName", "statement", "status", "reason", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "userId": { "type": "integer", "format": "int64" }, "penName": { "type": "string" }, "statement": { "type": "string" }, "status": { "type": "string" }, "reason": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" }, "decidedAt": { "type": ["string", "null"], "format": "date-time" }, "decidedByUserId": { "type": ["integer", "null"], "format": "int64" } } },
      "NullableAuthorApplication": { "oneOf": [{ "type": "null" }, { "$ref": "#/components/schemas/AuthorApplication" }] },
...
      "ProfileUpdateRequest": { "type": "object", "required": ["displayName"], "properties": { "displayName": { "type": "string", "minLength": 1, "maxLength": 1024 } }, "additionalProperties": false },
      "ProgressRequest": { "type": "object", "required": ["bookId", "chapterId", "offset"], "properties": { "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": "integer", "format": "int64" }, "offset": { "type": "integer", "minimum": 0 } }, "additionalProperties": false },
      "BookmarkRequest": { "type": "object", "required": ["chapterId", "offset"], "properties": { "chapterId": { "type": "integer", "format": "int64" }, "offset": { "type": "integer", "minimum": 0 }, "note": { "type": "string" } }, "additionalProperties": false },
      "CommentRequest": { "type": "object", "required": ["content"], "properties": { "chapterId": { "type": ["integer", "null"], "format": "int64" }, "content": { "type": "string", "minLength": 1, "maxLength": 4000 } }, "additionalProperties": false },
      "RatingRequest": { "type": "object", "required": ["rating"], "properties": { "rating": { "type": "integer" } }, "additionalProperties": false },
      "AmountRequest": { "type": "object", "required": ["amount"], "properties": { "amount": { "type": "integer" } }, "additionalProperties": false },
      "AuthorApplicationRequest": { "type": "object", "required": ["penName", "statement"], "properties": { "penName": { "type": "string", "minLength": 1, "maxLength": 128 }, "statement": { "type": "string", "minLength": 1, "maxLength": 4000 } }, "additionalProperties": false },
...
      "RedeemRequest": { "type": "object", "required": ["code"], "properties": { "code": { "type": "string", "minLength": 1 } }, "additionalProperties": false },
...
    }
  }
}